requirements.txt
beautifulsoup4
click
gitpython
html5lib
lxml
pandas
python-dotenv
requests
spotipy